requirement.txt
flask==3.1.0
transformers==4.46.3
torch==2.5.1
torchvision==0.20.1
pandas==2.1.3
nltk==3.9.1
google-cloud-language==2.15.1
python-dotenv==1.0.1
accelerate>=0.26.0
praw==7.8.1